Moved MapAuthorizer to test scope
